Fix chain for block rule
At least here the drop rule in the INPUT chain doesn't block any
traffic.
Could be because OpenVPN doesn't "create" the packets, but injects them
into the tun adapter. So it's like forwarding packets for the network
stack. Packets from tun interface are forwarded to the external
interface (eth0, or whatever).