Fixup block rule chain fix
Of course, chain has to be changed in .py, too.

Explanation of hotfix is bullshit.
INPUT chain is for packets directed to the host, so it is for protecting
the openvpn gateway itself.
FORWARD chain is for forwarded packets, which is what we are doing most
of the time.
So the best thing would probably be to make 2 block rules (INPUT and
FORWARD), but to me FORWARD chain seems more important.